Forgot to push this code, but this was from orrientation code refactoring on 9/26/23. More changes will be made here for refactoring and I will likely make a new project for code practices for easier implementation and universal utility.
