+++ src/index.html
@@ -15,7 +15,7 @@
             <div class="row">
                 <button class="green-button float-end m-1 ms-3 col-auto" id="addGoal"> Add A Goal + </button>
             </div>
-            <form class="goal-form visually-hidden" id="goalForm">
+            <form class="goal-form" id="goalForm">
                 <div class="row m-2">
                     <label class="col-2" for="title">Goal To Meet:</label>
                     <input class="col-lg-9 col-sm-8 col-7"
