--- src/js/general.js
@@ -1,6 +1,0 @@
-// CSS files
-import 'bootstrap/dist/css/bootstrap.min.css';
-import '../css/main.css';
-
-//Bootstrap JS File
-import 'bootstrap';
